EAB-105 Check for all boxes being empty and return an empty string in that case, move month converter to utils
